Modify Firebase token update due to deprecation
